added in CRUD for each controller: company, portfolio and stockviewAPI. not quite finished but adding in endpoints for later
